+++ burt-gui/src/config.js
@@ -37,7 +37,7 @@ const config = {
         createChatBotMessage("I'll assist you in reporting any problem with your app"),
         // createChatBotMessage("Please <b>start a stopwatch from scratch</b> to time yourself during" +
         //     " this conversation"),
-        createChatBotMessage("To start, please provide your <b>Participant ID</b>"),
+        createChatBotMessage("To start, please provide your <b>Participant Name</b>"),
     ],
     widgets: [
         {
