+++ burt-server/src/main/java/sealab/burt/server/StateVariable.java
@@ -3,7 +3,7 @@ package sealab.burt.server;
 public enum StateVariable {
 
     //general ones
-    NEXT_INTENTS, CURRENT_MESSAGE, SESSION_ID, PARTICIPANT_ID, PARTICIPANT_VALIDATED, PARTICIPANT_ASKED,
+    NEXT_INTENTS, CURRENT_MESSAGE, SESSION_ID, PARTICIPANT_ID, PARTICIPANT_NAME, PARTICIPANT_VALIDATED, PARTICIPANT_ASKED,
     CONFIRM_END_CONVERSATION, LAST_ACTION, ACTION_NEGATIVE_END_CONVERSATION, LAST_MESSAGE,
     MSG_NEGATIVE_END_CONVERSATION, CONFIRM_END_CONVERSATION_NEGATIVE,
